收藏实际价格数据库存放改成范围

+++ 51fbapi-biz/src/main/java/com/ald/fanbei/api/biz/service/AfuserCollectionService.java
@@ -40,6 +40,6 @@ public interface AfuserCollectionService {
 	
 	int deleteUserCollectionGoods(AfUserCollectionDo afUserCollectionDo);
 	
-	long getUserCollectionCountByGoodsId(AfUserCollectionDo afUserCollectionDo);
+    Integer getUserCollectionCountByGoodsId(AfUserCollectionDo afUserCollectionDo);
 
 }
