+++ 51fbapi-dal/src/main/java/com/ald/fanbei/api/dal/dao/AfUserCollectionDao.java
@@ -42,5 +42,5 @@ public interface AfUserCollectionDao {
 	
 	int deleteUserCollectionGoods(AfUserCollectionDo afUserCollectionDo);
 
-	long getUserCollectionCountByGoodsId(AfUserCollectionDo afUserCollectionDo);
+	Integer getUserCollectionCountByGoodsId(AfUserCollectionDo afUserCollectionDo);
 }
